Bulk namespace adjustment for appearance related models
