--- FashionSense/Framework/Models/AppearanceSync.cs
@@ -1,8 +1,0 @@
-﻿namespace FashionSense.Framework.Models
-{
-    public class AppearanceSync
-    {
-        public AppearanceContentPack.Type TargetAppearanceType { get; set; }
-        public AnimationModel.Type AnimationType { get; set; }
-    }
-}
