+++ FashionSense/Framework/Models/Appearances/AppearanceSync.cs
@@ -1,0 +1,8 @@
+﻿namespace FashionSense.Framework.Models.Appearances
+{
+    public class AppearanceSync
+    {
+        public AppearanceContentPack.Type TargetAppearanceType { get; set; }
+        public AnimationModel.Type AnimationType { get; set; }
+    }
+}
