+++ FashionSense/Framework/Models/Appearances/Generic/ConditionGroup.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace FashionSense.Framework.Models.Appearances.Generic
+{
+    public class ConditionGroup
+    {
+        public List<Condition> Conditions { get; set; }
+    }
+}
