+++ FashionSense/Framework/Models/Appearances/Generic/Position.cs
@@ -1,0 +1,8 @@
+﻿namespace FashionSense.Framework.Models.Appearances.Generic
+{
+    public class Position
+    {
+        public int X { get; set; }
+        public int Y { get; set; }
+    }
+}
