+++ FashionSense/Framework/Models/Appearances/Generic/Random/RandomRange.cs
@@ -1,0 +1,8 @@
+﻿namespace FashionSense.Framework.Models.Appearances.Generic.Random
+{
+    class RandomRange
+    {
+        public int Min { get; set; }
+        public int Max { get; set; }
+    }
+}
