+++ FashionSense/Framework/Models/Appearances/Generic/Size.cs
@@ -1,0 +1,8 @@
+﻿namespace FashionSense.Framework.Models.Appearances.Generic
+{
+    public class Size
+    {
+        public int Width { get; set; }
+        public int Length { get; set; }
+    }
+}
