+++ FashionSense/Framework/Models/Appearances/Hair/HairModel.cs
@@ -1,0 +1,13 @@
+﻿using FashionSense.Framework.Models.Appearances.Generic;
+
+namespace FashionSense.Framework.Models.Appearances.Hair
+{
+    public class HairModel : AppearanceModel
+    {
+        public Position HeadPosition { get; set; } = new Position() { X = 0, Y = 0 };
+        public Size HairSize { get; set; }
+
+        public override bool HideWhileSwimming { get; set; } = false;
+        public override bool HideWhileWearingBathingSuit { get; set; } = false;
+    }
+}
