--- FashionSense/Framework/Models/Generic/ConditionGroup.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace FashionSense.Framework.Models.Generic
-{
-    public class ConditionGroup
-    {
-        public List<Condition> Conditions { get; set; }
-    }
-}
