--- FashionSense/Framework/Models/Generic/Position.cs
@@ -1,8 +1,0 @@
-﻿namespace FashionSense.Framework.Models.Generic
-{
-    public class Position
-    {
-        public int X { get; set; }
-        public int Y { get; set; }
-    }
-}
