--- FashionSense/Framework/Models/Generic/Random/RandomRange.cs
@@ -1,8 +1,0 @@
-﻿namespace FashionSense.Framework.Models.Generic.Random
-{
-    class RandomRange
-    {
-        public int Min { get; set; }
-        public int Max { get; set; }
-    }
-}
