--- FashionSense/Framework/Models/Generic/Size.cs
@@ -1,8 +1,0 @@
-﻿namespace FashionSense.Framework.Models.Generic
-{
-    public class Size
-    {
-        public int Width { get; set; }
-        public int Length { get; set; }
-    }
-}
