+++ FashionSense/Framework/UI/SearchMenu.cs
@@ -1,11 +1,11 @@
-﻿using FashionSense.Framework.Models;
-using FashionSense.Framework.Models.Accessory;
-using FashionSense.Framework.Models.Hair;
-using FashionSense.Framework.Models.Hat;
-using FashionSense.Framework.Models.Pants;
-using FashionSense.Framework.Models.Shirt;
-using FashionSense.Framework.Models.Shoes;
-using FashionSense.Framework.Models.Sleeves;
+﻿using FashionSense.Framework.Models.Appearances;
+using FashionSense.Framework.Models.Appearances.Accessory;
+using FashionSense.Framework.Models.Appearances.Hair;
+using FashionSense.Framework.Models.Appearances.Hat;
+using FashionSense.Framework.Models.Appearances.Pants;
+using FashionSense.Framework.Models.Appearances.Shirt;
+using FashionSense.Framework.Models.Appearances.Shoes;
+using FashionSense.Framework.Models.Appearances.Sleeves;
 using FashionSense.Framework.Utilities;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
